+++ src/main/java/com/teamtea/eclipticseasons/mixin/compat/snowyspirit/MixinSnowySpiritGroundStatus.java
@@ -15,8 +15,7 @@ import org.spongepowered.asm.mixin.injection.At;
 public abstract class MixinSnowySpiritGroundStatus {
 
     @ModifyExpressionValue(at = {@At(value = "INVOKE", target = "Lnet/minecraft/world/level/Level;getBlockState(Lnet/minecraft/core/BlockPos;)Lnet/minecraft/world/level/block/state/BlockState;")},
-            method = {"computeFriction"},
-            remap = false)
+            method = {"computeFriction"})
     private static BlockState eclipticseasons$computeFriction(BlockState original,
                                                        @Local(argsOnly = true) Entity sled) {
         if (EclipticSeasonsApi.getInstance().isSnowyBlock(sled.level(), original, sled.getOnPos()))